requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
SQLAlchemy==2.0.11
psycopg2==2.9.6
pydantic==1.10.7
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
